requirements.txt
emo-platform-api-sdk